Consolidate headers into 1 for ease of consumption
