+++ libcassie/cassie_blob.cc
@@ -12,7 +12,7 @@
 #include <libcassandra/keyspace.h>
 
 #include "cassie.h"
-#include "cassie_blob.h"
+#include "cassie_private.h"
 
 namespace libcassie {
