+++ libcassie/include.am
@@ -7,12 +7,7 @@ EXTRA_DIST+= \
 
 nobase_include_HEADERS+= \
 			 libcassie/cassie.h \
-			 libcassie/cassie_types.h \
-			 libcassie/cassie_blob.h \
-			 libcassie/cassie_io_column.h \
-			 libcassie/cassie_io_super_column.h \
-			 libcassie/cassie_column.h \
-			 libcassie/cassie_super_column.h
+			 libcassie/cassie_private.h
 
 lib_LTLIBRARIES+= libcassie/libcassie.la
 libcassie_libcassie_la_CXXFLAGS= ${AM_CXXFLAGS}
